refatoração

+++ src/main/java/br/uniesp/si/techback/service/GeneroService.java
@@ -1,10 +1,11 @@
 package br.uniesp.si.techback.service;
 
 
+import br.uniesp.si.techback.domain.dto.request.GeneroRequestDTO;
 import br.uniesp.si.techback.exception.EntidadeNaoEncontradaException;
-import br.uniesp.si.techback.model.Genero;
-import br.uniesp.si.techback.model.dto.GeneroDTO;
+import br.uniesp.si.techback.domain.model.Genero;
 import br.uniesp.si.techback.repository.GeneroRepository;
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,35 +16,44 @@ import java.util.List;
 public class GeneroService {
     private GeneroRepository repository;
 
-    public Genero criar(GeneroDTO dto) {
-        Genero novoGenero = Genero.builder()
-                .descricao(dto.descricao())
-                .build();
-        repository.save(novoGenero);
-        return novoGenero;
-
-    }
-
     public Genero buscarPorId(Long id) {
-        Genero genero = repository.findById(id).orElseThrow(() -> new EntidadeNaoEncontradaException("Gênero não encontrado com o ID: " + id));
-        return genero;
+        return repository.findById(id)
+                .orElseThrow(() -> new EntidadeNaoEncontradaException("Gênero não encontrado com o ID: " + id));
     }
 
     public List<Genero> listar() {
         return repository.findAll();
     }
 
-    public void excluir(Long id) {
-        repository.deleteById(id);
+    @Transactional
+    public Genero salvar(GeneroRequestDTO dto) {
+        Genero novoGenero = Genero.builder()
+                .descricao(dto.descricao())
+                .build();
+        repository.save(novoGenero);
+        return novoGenero;
     }
 
-    public Genero atualizar(Long id, GeneroDTO dto) {
-        Genero genero = repository.findById(id).orElseThrow(() -> new EntidadeNaoEncontradaException("Gênero não encontrado com o ID: " + id));
+    @Transactional
+    public Genero atualizar(Long id, GeneroRequestDTO dto) {
+
+        if (!repository.existsById(id)) {
+            throw new EntidadeNaoEncontradaException("Gênero não encontrado com o ID: " + id);
+        }
+
         Genero generoEditado = Genero.builder()
                 .id(id)
                 .descricao(dto.descricao())
                 .build();
-        repository.save(generoEditado);
-        return generoEditado;
+        return repository.save(generoEditado);
+    }
+
+
+    @Transactional
+    public void excluir(Long id) {
+        if (!repository.existsById(id)) {
+            throw new EntidadeNaoEncontradaException("Genero não encontrado com o ID: " + id);
+        }
+        repository.deleteById(id);
     }
 }
